backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.135.1",
    "mysqlclient>=2.2.8",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "alembic>=1.18.4",
    "mypy>=1.20.0",
    "pytest>=9.0.2",
]
